requirements.txt
pypng==0.0.20
opencv-python==4.4.0.44
Pillow==8.0.1
